Task 1-7

+++ 03-strings/Strings/Task1/Program.cs
@@ -28,7 +28,7 @@ namespace Task1
                 sumWordLengths += word.Length;
             }
 
-            int avgWordLength = sumWordLengths / spltUserInput.Length;
+            double avgWordLength = (double)sumWordLengths / (double)spltUserInput.Length;
             Console.WriteLine(avgWordLength);
         }
     }
